// frontend-site-nextjs/src/types/types.ts
export interface ArticleResult {
  id: number;
  title: string;
  content: string;
  plainContent: string;
  likeCount: number;
  status: string | null;
  authorId: number | null;
  author: {
    id: number;
    firstName: string;
    lastName: string;
    avatar: string | null;
  };
  thumbnail: PhotosResult | null;
  categories: {
    id: number;
    title: string;
  }[];
  createdAt: Date;
  updatedAt: Date;
}
// ...
export interface PhotosResult {
  id: number;
  fullPath: string;
  path: string;
  createdAt: Date;
  updatedAt: Date;
}